+++ CHANGELOG.md
@@ -11,14 +11,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Options` menu is now visible
 - Dependency injection for much, much cleaner code
+- Brand new macro command engine
 
 ### Changed
 
 - Support for Dark Ages Client 7.41
+- Default window size is larger for modern displays
 - Refactor status window and updating of UI
 - Refactor client memory reading for 7.41
 - Optimize some graphics allocations
 - Removed the long-defunct `Chat` window menu
+- Removed the Debug / Logic Skeleton views
 - Increased font size
 - Updated to .NET Framework 4.8 (from 3.5)
 - Re-wrote Win32 window interop